change .js to .mjs. Tweak single post display. Add author to single post representation
