+++ feed/index.html
@@ -10,8 +10,8 @@
   <link href="https://fonts.googleapis.com/css2?family=Mulish:ital,wght@0,200..1000;1,200..1000&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="../dist/css/index.css">
   <script defer src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-  <script defer type="module" src="/js/posts.js"></script>
-  <script defer type="module" src="/js/createpost.js"></script>
+  <script defer type="module" src="/js/posts.mjs"></script>
+  <script defer type="module" src="/js/createpost.mjs"></script>
 </head>
 <body class="bg-login-body-green-white">
   <header>
