+++ feed/post.html
@@ -10,8 +10,8 @@
   <link href="https://fonts.googleapis.com/css2?family=Mulish:ital,wght@0,200..1000;1,200..1000&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="../dist/css/index.css">
   <script defer src="../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
-  <script defer src="../js/post.js" type="module"></script>
-  <script defer src="../js/api/single-post-fetch.js" type="module"></script>
+  <script defer src="../js/post.mjs" type="module"></script>
+  <script defer src="../js/api/single-post-fetch.mjs" type="module"></script>
 </head>
 <body class="bg-login-body-green-white">
   <header>
@@ -26,7 +26,7 @@
       </div>
       <div class="col-3 text-center pt-2">
         <a href="../profile/index.html">
-          <img src="../assets/img/profile-small.png" alt="Logo of original poster" class="img-fluid">
+          <img src="../assets/logo/profile-image-header-large.png" alt="Profile picture of logged in user" class="img-fluid">
         </a>
       </div>
     </div>
@@ -43,7 +43,7 @@
         </div>
         <div class="col-2">
           <a href="../profile/index.html">
-            <img src="../assets/img/profile-small.png" alt="Profile picture of logged in user">
+            <img src="../assets/logo/profile-image-header-large.png" alt="Profile picture of logged in user">
           </a>
         </div>
       </div>
