+++ index.html
@@ -9,7 +9,7 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Mulish:ital,wght@0,200..1000;1,200..1000&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="dist/css/index.css">
-  <script defer type="module" src="/js/loginform.js"></script>
+  <script defer type="module" src="/js/loginform.mjs"></script>
 </head>
 <body class="bg-login-body-green-white">
   <main>
