+++ js/api/posts-fetch.mjs
@@ -1,12 +1,10 @@
+import { headers } from "./constants.mjs";
+
 export async function fetchPosts(url, sortOrder = "desc", includeAuthor = true) {
   try {
-    const token = localStorage.getItem("accessToken");
     const getData = {
       method: "GET",
-      headers: {
-        'Content-Type': "application/json",
-        Authorization: `Bearer ${token}`,
-      },
+      headers: headers,
     };
 
     const urlWithSortOrderAndAuthor = `${url}?sortOrder=${sortOrder}&_author=${includeAuthor}`;
