+++ js/api/single-post-fetch.mjs
@@ -1,7 +1,4 @@
-
-
-
-import { apiBaseUrl, postsEndpoint } from "./constants.js"; 
+import { apiBaseUrl, postsEndpoint, headers } from "./constants.mjs"; 
 
 // Api call for specific post 
 
@@ -9,16 +6,9 @@ const queryString = document.location.search;
 const params = new URLSearchParams(queryString);
 const id = params.get("id");
 
-const token = localStorage.getItem("accessToken");
-
-const headers = {
-  'Content-Type': "application/json",
-  'Authorization': `Bearer ${token}`,
-};
-
 export async function getSpecificPost() {
   try {
-    const response = await fetch(`${apiBaseUrl}${postsEndpoint}${id}`, {
+    const response = await fetch(`${apiBaseUrl}${postsEndpoint}${id}?_author=true`, {
       method: "GET",
       headers: headers,
     });
