+++ profile/index.html
@@ -10,7 +10,7 @@
   <link href="https://fonts.googleapis.com/css2?family=Mulish:ital,wght@0,200..1000;1,200..1000&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="../dist/css/index.css">
   <script defer src="../js/api/profile-posts-fetch.js" type="module"></script>
-  <script defer src="../js/profile.js" type="module"></script>
+  <script defer src="../js/profile.mjs" type="module"></script>
   <script defer src="../js/profileName.js" type="module"></script>
   <script defer src="../js/logOut.js"></script>
   <script src="https://kit.fontawesome.com/99f531cd0d.js" crossorigin="anonymous"></script>
